requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
pillow-heif==0.16.0
python-multipart==0.0.9
